# [JRuby] Optimize `scan()`: Use `strBL.getBegin() + curr` instead of `currPtr()`